chore: moved mail-limits to security domain
