+++ src/modules/security/mail-limit/mail-limit.domain.ts
@@ -1,5 +1,5 @@
 import { Time } from '../../../lib/time';
-import { MailTypes } from '../mailTypes';
+import { MailTypes } from './mailTypes';
 import { MailLimitModelAttributes } from './mail-limit.model';
 
 export class MailLimit implements MailLimitModelAttributes {
