+++ src/modules/security/mail-limit/mail-limit.repository.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common';
 
 import { MailLimitModel, MailLimitModelAttributes } from './mail-limit.model';
 import { MailLimit } from './mail-limit.domain';
-import { MailTypes } from '../mailTypes';
+import { MailTypes } from './mailTypes';
 
 @Injectable()
 export class SequelizeMailLimitRepository {
